+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Add `key` to `CrowdfundingCard` video sources.
 - Fix: Remove console error from `Header` component `Centered`.
 
 ## [2.86.0] - 2020-08-17