+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `Video` component to display videos on `RewardCard`.
+- Fix: Add `key` to `CrowdfundingCard` video sources.
+- Fix: Remove console error from `Header` component `Centered`.
 
 ## [2.86.0] - 2020-08-17