+++ CHANGELOG.md
@@ -4,9 +4,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `Video` component to display videos on `RewardCard`.
-- Fix: Add `key` to `CrowdfundingCard` video sources.
-- Fix: Remove console error from `Header` component `Centered`.
+## [2.87.0] - 2020-08-18
+
+Feature:
+- Add `Video` component to display videos on `RewardCard`.
+
+Fixes:
+- Add `key` to `CrowdfundingCard` video sources.
+- Remove console error from `Header` component `Centered`.
 
 ## [2.86.0] - 2020-08-17